Move SignIn and SignUp into auth directory
